requirements.txt
python-dotenv
requests
pandas
tqdm
matplotlib
numpy
seaborn
python-dotenv